+++ src/Either.ts
@@ -1197,10 +1197,6 @@ export const ChainRec: ChainRec2<URI> = {
  */
 export const MonadThrow: MonadThrow2<URI> = {
   URI,
-  map: map_,
-  ap: ap_,
-  of,
-  chain: chain_,
   throwError: throwError
 }